First commit
Create new Elixir project with Mix:

    mix new --sup twittex

The projects depends on `oauther` and `oauth2` Hex packages. Both are
required to implement OAuth1.0 and OAuth2.0 protocols.

Twitter provides application-only authentication based on the `Client
Credential Grant` flow of the OAuth2.0 specification.

With application-only authentication you don’t have the context of an
authenticated user and this means that any request to API for endpoints
that require user context, such as posting tweets, will not work.

Twitter also supports the `xAuth` OAuth1.0 extension to exchange a
username and password for an OAuth access token.