Refactor last commit
Twitter.Client.stream/2 now uses a pub/sub pattern using two processes.
One is a spawned process (not linked to the calling process), the other
is a GenEvent event manager.

    iex(1)> stream = Twittex.Client.stream "hello"
    %GenEvent.Stream{manager: #PID<0.169.0>, timeout: :infinity}
    iex(2)> GenEvent.stop stream.manager
    :ok

We use GenEvent.stream/1 to create a Stream from the incoming events.
To kill both, the publisher and the subscriber, you only have to stop
the GenEvent. This can be done using the :manager reference from the
returned stream. Would be nice if the HTTP connection would be close
when the spawned process dies. Not sure if that is true.