Modify Twittex.Client.stream/2 to close connection
The Twittex.Client.stream/2 function returns a %GenEvent.Stream{}.
The Twitter Stream API does not close the HTTP connection at any time
because the stream is endless. This means that the user must close the
connection manually. This is done by stopping the GenEvent.

1) The spawned process transforming the incoming chunks to JSON payloads
will be killed.
2) The :hackney process keeping the connection open and reading incoming
data will also be stopped.